+++ src/standard/event_list.rs
@@ -48,7 +48,7 @@ impl EventList {
         match self.event_list_core.get_clone(index) {
             Some(ev_res) => {
                 match ev_res {
-                    Ok(ev) => Some(Event::from_event_core(ev)),
+                    Ok(ev) => Some(Event::new(ev)),
                     Err(_) => None,
                 }
             },
@@ -61,7 +61,7 @@ impl EventList {
         match self.event_list_core.last_clone() {
             Some(ev_res) => {
                 match ev_res {
-                    Ok(ev) => Some(Event::from_event_core(ev)),
+                    Ok(ev) => Some(Event::new(ev)),
                     Err(_) => None,
                 }
             },