+++ src/standard/queue.rs
@@ -1,7 +1,7 @@
 //! An OpenCL command queue.
 
 use core::{self, CommandQueue as CommandQueueCore, DeviceId as DeviceIdCore, Context as ContextCore};
-use super::Context;
+use standard::{Context, Device};
 
 /// A command queue.
 ///
@@ -18,6 +18,38 @@ pub struct Queue {
 }
 
 impl Queue {
+    /// Returns a new Queue on the device specified by `device`. 
+    ///
+    /// Not specifying `device` will default to the first available device
+    /// associated with `context`.
+    ///
+    /// [FIXME]: Return result.
+    pub fn new(context: &Context, device: Option<Device>) -> Queue {
+        // let device_idxs = match device_idx {
+        //     Some(idx) => vec![idx],
+        //     None => Vec::with_capacity(0),
+        // };
+
+        // let device_ids_core = context.resolve_device_idxs(&device_idxs);
+        // assert!(device_ids_core.len() == 1, "Queue::new_by_device_index: Error resolving device ids.");
+        // let device_id_core = device_ids_core[0].clone();
+
+        let device_id_core = match device {
+            Some(d) => d.as_core().clone(),
+            None => context.get_device_by_index(0).as_core().clone(),
+        };
+
+        let obj_core = core::create_command_queue(context.core_as_ref(), &device_id_core)
+            .expect("[FIXME: TEMPORARY]: Queue::new_by_device_index():"); 
+
+        Queue {
+            obj_core: obj_core,
+            context_obj_core: context.core_as_ref().clone(),
+            device_id_core: device_id_core, 
+        }
+    }
+
+
     /// Returns a new Queue on the device specified by `device_idx`. 
     ///
     /// 'device_idx` refers to a index in the list of devices generated when creating
@@ -27,25 +59,26 @@ impl Queue {
     /// the documentation for `Context` for more information.
     /// 
     /// [FIXME]: Return result.
-    pub fn new(context: &Context, device_idx: Option<usize>) -> Queue {
+    pub fn new_by_device_index(context: &Context, device_idx: Option<usize>) -> Queue {
         let device_idxs = match device_idx {
             Some(idx) => vec![idx],
             None => Vec::with_capacity(0),
         };
 
         let device_ids_core = context.resolve_device_idxs(&device_idxs);
-        assert!(device_ids_core.len() == 1, "Queue::new: Error resolving device ids.");
+        assert!(device_ids_core.len() == 1, "Queue::new_by_device_index: Error resolving device ids.");
         let device_id_core = device_ids_core[0].clone();
 
         let obj_core = core::create_command_queue(context.core_as_ref(), &device_id_core)
-            .expect("[FIXME: TEMPORARY]: Queue::new():"); 
+            .expect("[FIXME: TEMPORARY]: Queue::new_by_device_index():"); 
 
         Queue {
             obj_core: obj_core,
             context_obj_core: context.core_as_ref().clone(),
             device_id_core: device_id_core, 
         }
-    }   
+    }  
+
 
     /// Blocks until all commands in this queue have completed.
     pub fn finish(&self) {